% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_vus.R
\name{add_info}
\alias{add_info}
\title{Count the number of "supporting", "moderate", "strong" and "very strong" strengths of evidence for pathogenicity}
\usage{
add_info(data, classification_col)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{classification_col}{The column name for variant classification (str). Variants should be classified into five distinct categories: "P," "LP," "B," "LB," and "VUS."}
}
\value{
A new DataFrame that includes the input data and four new columns, these four columns count the number of different pathogenic evidence strengths for each variant, which can be used for further categorization
}
\description{
Count the number of "supporting", "moderate", "strong" and "very strong" strengths of evidence for pathogenicity
}
\examples{
data("ClinGen_dataset")
ClinGen_dataset <- add_info(ClinGen_dataset, "Assertion")

}
