% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_classification.R
\name{BCF}
\alias{BCF}
\title{Classifying variants into five distinct categories according to the Bayesian classification framework}
\usage{
BCF(data, evidence_col, prior_p, op_vs)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{evidence_col}{The column name for ACMG evidence(str)}

\item{prior_p}{The prior probability of pathogenicity (proportion of P/LP variants in a set of variants)}

\item{op_vs}{Odds path of "Very String"}
}
\value{
A new DataFrame that incorporates the input data and the results of variant classification
}
\description{
Classifying variants into five distinct categories according to the Bayesian classification framework
}
\examples{
data("VCI_data")
BCF(VCI_data, "Applied Evidence Codes (Met)", 0.1, 350)

}
