% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ClinGen_dataset}
\alias{ClinGen_dataset}
\title{The ClinGen Curated Variants dataset}
\format{
A data frame with 6768 rows and 20 variables:
\describe{
\item{#Variation}{Variation, in HGVSc}
\item{ClinVar Variation Id}{ClinVar Variation ID}
\item{Allele Registry Id}{ClinGen Allele Registry ID}
\item{HGVS Expressions}{HGVS Expressions in ClinVar}
\item{HGNC Gene Symbol}{Gene Symbol}
\item{Disease}{Variant related disease}
\item{Mondo Id}{Mondo Disease Ontology ID}
\item{Mode of Inheritance}{Genetic Inheritance pattern}
\item{Assertion}{Variant Classification}
\item{Applied Evidence Codes (Met)}{Criteria, represent following the SVI's recommendations}
\item{Applied Evidence Codes (Not Met)}{Criteria not used}
\item{Summary of interpretation}{Detailed information for each applied criteria}
\item{PubMed Articles}{PubMed ID}
\item{Expert Panel}{The name of variant curation expert panel}
\item{Guideline}{Links of specific guidelines}
\item{Approval Date}{Approval Date}
\item{Published Date}{Published Date}
\item{Retracted}{Retracted, in logical}
\item{Evidence Repo Link}{Evidence Repo Link}
\item{Uuid}{ID}
...
}
}
\source{
\url{https://erepo.clinicalgenome.org/evrepo/}
}
\usage{
ClinGen_dataset
}
\description{
This dataset encompasses classification summaries for 6,768 curated variants across 74 diseases, including 1850 P, 1463 LP, 679 LB, 775 B, and 2001 US variants.
}
\keyword{datasets}
