% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ClinVar_2019_dataset}
\alias{ClinVar_2019_dataset}
\title{The ClinVar 2019 dataset compiled by Pejaver et al.}
\format{
A data frame with 11834 rows and 27 variables:
\describe{
\item{hg19_chr}{Chromosome}
\item{hg19_pos(1-based)}{Position}
\item{ref}{Reference allele}
\item{alt}{Alternative allele}
\item{rs_dbSNP151}{rsID}
\item{genename}{Gene name}
\item{Ensembl_geneid}{GeneID}
\item{Ensembl_transcriptid}{TranscriptID}
\item{Ensembl_proteinid}{ProteinID}
\item{Uniprot_acc}{Uniprot Accession}
\item{Uniprot_entry}{UniProt entry name}
\item{aavar}{AA change}
\item{clnsig}{ClinVar Significance}
\item{MAF}{Minor allele frequency}
\item{SIFT_score}{SIFT score}
\item{FATHMM_score}{FATHMM score}
\item{VEST4_score}{VEST4 score}
\item{REVEL_score}{REVEL score}
\item{GERP++_RS}{GERP++ score}
\item{phyloP100way_vertebrate}{phyloP score}
\item{EA_1.0}{EA score}
\item{BayesDel_nsfp33a_noAF}{BayesDel score}
\item{MutPred2.0_score}{MutPred score}
\item{CADDv1.6_PHRED}{CADD score}
\item{pph2_prob}{pph2 score}
\item{MPC_score}{MPC score}
\item{PrimateAI_score}{PrimateAI score}
...
}
}
\source{
\url{https://zenodo.org/records/8347415}
}
\usage{
ClinVar_2019_dataset
}
\description{
11,834 variants (2,787 P/LP and 6,327 B/LB variants) from 1,914 genes are included in this dataset.
}
\keyword{datasets}
