% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_LR.R
\name{discrete_cutoff}
\alias{discrete_cutoff}
\title{Introducing columns to assess if the observed value is above (1) or below (0) a tested cutoff. A value of 1 indicates being above the tested cutoff, while 0 indicates being below the tested cutoff}
\usage{
discrete_cutoff(data, feature, range = NULL, criteria = NULL)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{feature}{The column name that requires testing for optimizing the thresholds}

\item{range}{Evaluated intervals}

\item{criteria}{ACMG/AMP guidelines criteria (str)}
}
\value{
A fresh DataFrame incorporating the input data with additional column
}
\description{
Introducing columns to assess if the observed value is above (1) or below (0) a tested cutoff. A value of 1 indicates being above the tested cutoff, while 0 indicates being below the tested cutoff
}
\examples{
data("ClinGen_dataset")
discrete_cutoff(ClinGen_dataset, "Applied Evidence Codes (Met)", criteria = "PM2")

}
