% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalVariables.R
\name{.onLoad}
\alias{.onLoad}
\title{Define the global Variables}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{lib name}

\item{pkgname}{package name}
}
\value{
global variables
}
\description{
Define the global Variables
}
\examples{
#null

}
