% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_op.R
\name{evaluation_variant}
\alias{evaluation_variant}
\title{Indicating candidate variants for re-evaluation}
\usage{
evaluation_variant(data, evidence_col, r_threshold = 0.6)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{evidence_col}{The column name for the ACMG/AMP evidence (str)}

\item{r_threshold}{The threshold for evidence correlation (int)}
}
\value{
evidence correlation
}
\description{
Indicating candidate variants for re-evaluation
}
\examples{
data("ClinGen_dataset")
evaluation_variant(ClinGen_dataset, "Applied Evidence Codes (Met)")

}
