% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{evidence_corplot}
\alias{evidence_corplot}
\title{Visualize the correlation of ACMG/AMP evidence}
\usage{
evidence_corplot(data, evidence_col)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{evidence_col}{The column name for the ACMG/AMP evidence (str)}
}
\value{
Figures
}
\description{
Visualize the correlation of ACMG/AMP evidence
}
\examples{
data("ClinGen_dataset")
evidence_corplot(ClinGen_dataset, "Applied Evidence Codes (Met)")

}
