% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{heatmap_LR}
\alias{heatmap_LR}
\title{Visualize the results of LR+ for each evaluated cutoff}
\usage{
heatmap_LR(data, op_list)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{op_list}{A list of odds path corresponding to each level of evidence strength}
}
\value{
Figures
}
\description{
Visualize the results of LR+ for each evaluated cutoff
}
\examples{
data("LR_result")
op_list <- c(2.11, 4.46, 19.90, 397)
heatmap_LR(LR_result, op_list)

}
