% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_locallr.R
\name{local_bootstrapped_lr}
\alias{local_bootstrapped_lr}
\title{The one-sided 95\% confidence bound for each estimated lr+  was determined through bootstrapping iterations, enabling the assessment of evidence strength.}
\usage{
local_bootstrapped_lr(
  input_data,
  feature,
  alpha,
  bootstrap,
  minpoints,
  increment,
  output_dir
)
}
\arguments{
\item{input_data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{feature}{The column name that requires testing for optimizing the thresholds}

\item{alpha}{Prior probability}

\item{bootstrap}{The number of bootstrapping iterations}

\item{minpoints}{The number of at least pathogenic and non-pathogenic variants}

\item{increment}{Sliding window}

\item{output_dir}{Output directory}
}
\value{
The posterior probability values for each bootstrap iteration
}
\description{
The one-sided 95\% confidence bound for each estimated lr+  was determined through bootstrapping iterations, enabling the assessment of evidence strength.
}
\examples{
data("ClinVar2020_AJHG_Pejaver_data")
ClinVar2020_AJHG_Pejaver_data <- add_info(ClinVar2020_AJHG_Pejaver_data, "clnsig")
local_bootstrapped_lr(ClinVar2020_AJHG_Pejaver_data, "PrimateAI_score", 0.1, 10, 100, 0.1, "test_dir")

}
