% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_locallr.R
\name{local_lr}
\alias{local_lr}
\title{Calculating the local positive likelihood ratio (lr+) value, which is applicable to continuous evidence proposed by Pejaver et al.
First, all unique tested cutoff values were sorted, then each value was positioned at the center of a sliding window.
The posterior probability was calculated for each tested cutoff value within the interval, considering a minimum of selected pathogenic and non-pathogenic variants.}
\usage{
local_lr(input_data, feature, direction, alpha, minpoints, increment)
}
\arguments{
\item{input_data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{feature}{The column name that requires testing for optimizing the thresholds}

\item{direction}{The direction of evidence pathogenic (Pathogenic or Benign)}

\item{alpha}{Prior probability}

\item{minpoints}{The number of at least pathogenic and non-pathogenic variants}

\item{increment}{Sliding window}
}
\value{
The posterior probability value for each tested cutoff
}
\description{
Calculating the local positive likelihood ratio (lr+) value, which is applicable to continuous evidence proposed by Pejaver et al.
First, all unique tested cutoff values were sorted, then each value was positioned at the center of a sliding window.
The posterior probability was calculated for each tested cutoff value within the interval, considering a minimum of selected pathogenic and non-pathogenic variants.
}
\examples{
\dontrun{
data("ClinVar_2019_dataset")
data <- add_info(ClinVar_2019_dataset, "clnsig")
local_lr(data, "PrimateAI_score", "Pathogenic",0.0441, 100, 0.01)
}

}
