% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_LR.R
\name{LR}
\alias{LR}
\title{Calculating positive likelihood ratio (LR) for each tested cutoff (for discrete cutoffs)
For each cutoff, true positive (TP, the number of P/LP variants above a tested cutoff),
false positive (FP, the number of BL-VUS/B/LB variants above a tested cutoff),
true negative (TN, the number of BL-VUS/B/LB variants below a tested cutoff),
and false negative (FN, the number of P/LP variants below a tested cutoff) were estimated.
Subsequently, LR+, overall accuracy, true positive rate (sensitivity), true negative rate (specificity), positive predictive value (PPV), negative predictive value (NPV), and F1 score were calculated.
Estimates of 95\% CI of LR+ were generated using bootstrapping in the R package, bootLR.}
\usage{
LR(data, start, end)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{start}{The beginning column index of the evaluated cutoffs}

\item{end}{The concluding column index of evaluated cutoffs}
}
\value{
A DataFrame comprising the evaluation metrics for each assessed cutoff
}
\description{
Calculating positive likelihood ratio (LR) for each tested cutoff (for discrete cutoffs)
For each cutoff, true positive (TP, the number of P/LP variants above a tested cutoff),
false positive (FP, the number of BL-VUS/B/LB variants above a tested cutoff),
true negative (TN, the number of BL-VUS/B/LB variants below a tested cutoff),
and false negative (FN, the number of P/LP variants below a tested cutoff) were estimated.
Subsequently, LR+, overall accuracy, true positive rate (sensitivity), true negative rate (specificity), positive predictive value (PPV), negative predictive value (NPV), and F1 score were calculated.
Estimates of 95\% CI of LR+ were generated using bootstrapping in the R package, bootLR.
}
\examples{
\dontrun{
data("ClinGen_dataset")
data <- add_info(ClinGen_dataset, "Assertion")
data <- VUS_classify(data, "Assertion", "Applied Evidence Codes (Met)")
#data <- data[data$`Applied Evidence Codes (Met)`!="",]
all_evidence <- unlist(str_replace_all(data$`Applied Evidence Codes (Met)`," ", ""))
split_evidence <- strsplit(all_evidence, ",")
unique_evidence <- unique(unlist(split_evidence))
P_evidence<-grep("^P", unique_evidence, value = TRUE)
library(dplyr)
truth_set <- filter(data,VUS_class \%in\% c("IceCold","Cold","Cool",""))
for(i in P_evidence){
  truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = i)
}
LR_result<-LR(truth_set, 28, 72)
rownames(LR_result)<-LR_result[,1]
LR_result<-LR_result[,-1]
name_evidence<-rownames(LR_result)
LR_result<-data.frame(lapply(LR_result,as.numeric))
rownames(LR_result)<-name_evidence
LR_result<-LR_result[c(-1,-2,-4,-5,-6,-7,-8,-10,-11,-12,-14,-17,-18,-19,-20,-21,-22,-24,-25,-26),]
LR_result<-LR_result[c(2,4,6,1,3,5),]
}

}
