% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lr_CI_result}
\alias{lr_CI_result}
\title{Local posterior probability and one-sided 95\% confidence intervals of the local posterior probability for each unique PrimateAI score in the ClinVar 2019 dataset}
\format{
A data frame with 8596 rows and 3 variables:
\describe{
\item{test_cutoff}{Each PrimateAI score}
\item{Posterior}{Posterior probability}
\item{Posterior1}{The 95\% CI lower boundry of posterior probability}
...
}
}
\source{
ClinVar_2019_dataset
}
\usage{
lr_CI_result
}
\description{
Local posterior probability and one-sided 95\% confidence intervals of the local posterior probability for each unique PrimateAI score in the ClinVar 2019 dataset
}
\keyword{datasets}
