% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LR_result}
\alias{LR_result}
\title{Evaluation metrics and positive likelihood ratio for PM2 and PM2_Supporting derived from the ClinGen Curated Variants dataset}
\format{
A data frame with 8 rows and 20 variables:
\describe{
\item{TP}{True positive}
\item{FN}{False negative}
\item{FP}{False positive}
\item{TN}{True negative}
\item{Accuracy}{(TP+TN)/Total}
\item{PPV}{Positive predictive values}
\item{NPV}{Negative predictive values}
\item{FNR}{False negative rate}
\item{FPR}{False positive rate}
\item{FOR}{False omission rate}
\item{FDR}{False discovery rate}
\item{F1}{F1 score}
\item{Sensitivity}{True positive rate}
\item{Specificity}{True negative rate}
\item{posLR}{Positive likelihood ratio}
\item{posLR_LB}{The 95\% CI lower boundry of posLR}
\item{posLR_UB}{The 95\% CI upper boundry of posLR}
\item{negLR}{Negative likelihood ratio}
\item{negLR_LB}{The 95\% CI lower boundry of negLR}
\item{negLR_UB}{The 95\% CI upper boundry of negLR}
...
}
}
\source{
ClinGen_dataset
}
\usage{
LR_result
}
\description{
Evaluation metrics and positive likelihood ratio for PM2 and PM2_Supporting derived from the ClinGen Curated Variants dataset
}
\keyword{datasets}
