% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{multi_plot}
\alias{multi_plot}
\title{Visualize the distribution of variants}
\usage{
multi_plot(data, classification_col, gene_col, consequence_col = NULL)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{classification_col}{The column name for variant classification (str)}

\item{gene_col}{The column name for the gene where the variant is located(str)}

\item{consequence_col}{The column name for the annotation results of variant consequences(str)}
}
\value{
Figures
}
\description{
Visualize the distribution of variants
}
\examples{
data("VCI_data")
VCI_data <- add_info(VCI_data, "Assertion")
VCI_data <- VUS_classify(VCI_data, "Assertion", "Applied Evidence Codes (Met)")
multi_plot(VCI_data, "Assertion", "HGNC Gene Symbol")

}
