% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{plot_lr}
\alias{plot_lr}
\title{Generate plots depicting the results of lr+ for each tested cutoff}
\usage{
plot_lr(data, direction, postp_list)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{direction}{The direction of evidence pathogenic (Pathogenic or Benign)}

\item{postp_list}{A list of posterior probability corresponding to each level of evidence strength}
}
\value{
Figures
}
\description{
Generate plots depicting the results of lr+ for each tested cutoff
}
\examples{
data("lr_CI_result")
# data <- add_info(ClinVar_2019_dataset, "clnsig")
# local_bootstrapped_lr(data, "PrimateAI_score", 0.0441, 10000, 100, 0.01, "test_dir")
postp_list <- c(0.100, 0.211, 0.608, 0.981)
# lr_CI_result <- lr_CI(30, "test_dir")
plot_lr(lr_CI_result, "Pathogenic", postp_list)

}
