% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_classification.R
\name{Point_Classification}
\alias{Point_Classification}
\title{Classifying variants into five distinct categories according to the scaled point system.}
\usage{
Point_Classification(data, evidence_col)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{evidence_col}{The column name for ACMG evidence(str)}
}
\value{
A new DataFrame that incorporates the input data and the results of variant classification
}
\description{
Classifying variants into five distinct categories according to the scaled point system.
}
\examples{
\dontrun{
data("VCI_data")
Point_Classification(VCI_data, "Applied Evidence Codes (Met)")
}

}
