% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VCI_data}
\alias{VCI_data}
\title{Clinical variant classification in ClinGen Variant Curation Interface}
\format{
A data frame with 5724 rows and 20 variables:
\describe{
\item{#Variation}{Variation, in HGVSc}
\item{ClinVar Variation Id}{ClinVar Variation ID}
\item{Allele Registry Id}{ClinGen Allele Registry ID}
\item{HGVS Expressions}{HGVS Expressions in ClinVar}
\item{HGNC Gene Symbol}{Gene Symbol}
\item{Disease}{Variant related disease}
\item{Mondo Id}{Mondo Disease Ontology ID}
\item{Mode of Inheritance}{Genetic Inheritance pattern}
\item{Assertion}{Variant Classification}
\item{Applied Evidence Codes (Met)}{Criteria, represent following the SVI's recommendations}
\item{Applied Evidence Codes (Not Met)}{Criteria not used}
\item{Summary of interpretation}{Detailed information for each applied criteria}
\item{PubMed Articles}{PubMed ID}
\item{Expert Panel}{The name of variant curation expert panel}
\item{Guideline}{Links of specific guidelines}
\item{Approval Date}{Approval Date}
\item{Published Date}{Published Date}
\item{Retracted}{Retracted, in logical}
\item{Evidence Repo Link}{Evidence Repo Link}
\item{Uuid}{ID}
...
}
}
\source{
\url{https://erepo.clinicalgenome.org/evrepo/}
}
\usage{
VCI_data
}
\description{
A dataset containing the curated 5724 variants by ClinGen.
}
\keyword{datasets}
