% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_op.R
\name{auto_select_postp}
\alias{auto_select_postp}
\title{Automatic definition of posterior probability and positive likelihood ratio values
for different strengths of evidence}
\usage{
auto_select_postp(prior_probability)
}
\arguments{
\item{prior_probability}{The prior probability of pathogenicity (proportion of P/LP variants in a set of variants)}
}
\value{
Selected odds path for each evidence level, combined odds path and posterior probability of 17 combination rules
}
\description{
Automatic definition of posterior probability and positive likelihood ratio values
for different strengths of evidence
}
\examples{
auto_select_postp(0.1)

}
