% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_locallr.R
\name{get_lr_threshold}
\alias{get_lr_threshold}
\title{Establish the thresholds for each level of evidence strength}
\usage{
get_lr_threshold(postp_list, discountonesided, bootstrap, dir)
}
\arguments{
\item{postp_list}{A list of posterior probability corresponding to each level of evidence strength}

\item{discountonesided}{The one-sided confidence intervals}

\item{bootstrap}{The number of bootstrapping iterations}

\item{dir}{The directory containing the results of bootstrapping}
}
\value{
A list of optimized thresholds
}
\description{
Establish the thresholds for each level of evidence strength
}
\examples{
\dontrun{
data("ClinVar2020_AJHG_Pejaver_data")
data <- add_info(ClinVar2020_AJHG_Pejaver_data, "clnsig")
local_bootstrapped_lr(data, "PrimateAI_score", 0.1, 10, 100, 0.1, "test_dir")
postp_list <- c(0.1778350, 0.3129676, 0.6689245, 0.9754584)
get_lr_threshold(postp_list, 0.05, 10, "test_dir")
}

}
