% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_LR.R
\name{LR}
\alias{LR}
\title{Calculating positive likelihood ratio (LR) for each tested cutoff (for discrete cutoffs)
For each cutoff, true positive (TP, the number of P/LP variants above a tested cutoff),
false positive (FP, the number of BL-VUS/B/LB variants above a tested cutoff),
true negative (TN, the number of BL-VUS/B/LB variants below a tested cutoff),
and false negative (FN, the number of P/LP variants below a tested cutoff) were estimated.
Subsequently, LR+, overall accuracy, true positive rate (sensitivity), true negative rate (specificity), positive predictive value (PPV), negative predictive value (NPV), and F1 score were calculated.
Estimates of 95\% CI of LR+ were generated using bootstrapping in the R package, bootLR.}
\usage{
LR(data, start, end)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{start}{The beginning column index of the evaluated cutoffs}

\item{end}{The concluding column index of evaluated cutoffs}
}
\value{
A DataFrame comprising the evaluation metrics for each assessed cutoff
}
\description{
Calculating positive likelihood ratio (LR) for each tested cutoff (for discrete cutoffs)
For each cutoff, true positive (TP, the number of P/LP variants above a tested cutoff),
false positive (FP, the number of BL-VUS/B/LB variants above a tested cutoff),
true negative (TN, the number of BL-VUS/B/LB variants below a tested cutoff),
and false negative (FN, the number of P/LP variants below a tested cutoff) were estimated.
Subsequently, LR+, overall accuracy, true positive rate (sensitivity), true negative rate (specificity), positive predictive value (PPV), negative predictive value (NPV), and F1 score were calculated.
Estimates of 95\% CI of LR+ were generated using bootstrapping in the R package, bootLR.
}
\examples{
data("VCI_data")
VCI_data <- add_info(VCI_data, "Assertion")
VCI_data <- VUS_classify(VCI_data, "Assertion", "Applied Evidence Codes (Met)")
truth_set <- VCI_data[VCI_data$VUS_class != "Hot" & VCI_data$VUS_class != "Warm" & VCI_data$VUS_class != "Tepid" & VCI_data$VUS_class != "NA", ]
truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = "PM2")
truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = "PP3")
truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = "PP1")
truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = "PVS1")
truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = "PS1")
truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = "PS2")
truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = "PM3")
truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = "PM4")
truth_set <- discrete_cutoff(truth_set, "Applied Evidence Codes (Met)", criteria = "PM5")
LR(truth_set, 28, 36)

}
