% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_locallr.R
\name{lr_CI}
\alias{lr_CI}
\title{Merging the results from bootstrap}
\usage{
lr_CI(bootstrap, dir)
}
\arguments{
\item{bootstrap}{The number of bootstrapping iterations}

\item{dir}{The directory containing the results of bootstrapping}
}
\value{
A DataFrame containing posterior probabilities and the 95\% confidence interval lower bounds of posterior probabilities for each cutoff
}
\description{
Merging the results from bootstrap
}
\examples{
\dontrun{
data("ClinVar_2019_dataset")
data <- add_info(ClinVar_2019_dataset, "clnsig")
local_bootstrapped_lr(data, "PrimateAI_score", 0.0441, 10000, 100, 0.01, "test_dir")
lr_CI_result <- lr_CI(10000, "test_dir")
}

}
