% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lr_CI_result}
\alias{lr_CI_result}
\title{locallr results for PrimateAI_score in ClinVar2020_AJHG_Pejaver_data}
\format{
A data frame with 8586 rows and 3 variables:
\describe{
\item{test_cutoff}{Each PrimateAI score}
\item{Posterior}{Posterior probability}
\item{Posterior1}{The 95\% CI lower boundry of posterior probability}
...
}
}
\source{
ClinVar2020_AJHG_Pejaver_data
}
\usage{
lr_CI_result
}
\description{
locallr results for PrimateAI_score in ClinVar2020_AJHG_Pejaver_data
}
\keyword{datasets}
