% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{plot_lr}
\alias{plot_lr}
\title{Generate plots depicting the results of lr+ for each tested cutoff}
\usage{
plot_lr(data, postp_list)
}
\arguments{
\item{data}{DataFrame comprising fundamental variant information, evidence labeling, and classification details}

\item{postp_list}{A list of posterior probability corresponding to each level of evidence strength}
}
\value{
Figures
}
\description{
Generate plots depicting the results of lr+ for each tested cutoff
}
\examples{
data("lr_CI_result")
# ClinVar2020_AJHG_Pejaver_data <- add_info(ClinVar2020_AJHG_Pejaver_data, "clnsig")
# local_bootstrapped_lr(ClinVar2020_AJHG_Pejaver_data, "PrimateAI_score", 0.1, 30, 100, 0.1, "test_dir")
postp_list <- c(0.1778350, 0.3129676, 0.6689245, 0.9754584)
# lr_CI_result <- lr_CI(30, "test_dir")
plot_lr(lr_CI_result, postp_list)

}
